scripts/requirements.txt
# Big Island VR 3.0 - Python Dependencies
# Install with: pip install -r requirements.txt

# Core
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
requests>=2.31.0

# Depth Estimation (Depth Anything v2)
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
huggingface-hub>=0.19.0
safetensors>=0.4.0

# Image Processing & Stitching
scikit-image>=0.21.0
scipy>=1.11.0

# AI Upscaling (Real-ESRGAN)
realesrgan>=0.3.0
basicsr>=1.4.2

# Video Generation
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# GPU Acceleration (optional - install separately if needed)
# For CUDA: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For MPS (Apple Silicon): pip install torch torchvision

# Progress & Utilities
tqdm>=4.66.0
rich>=13.6.0
